+++ driver-core/src/test/java/com/datastax/driver/core/LoadBalancingPolicyTest.java
@@ -24,6 +24,7 @@ import org.testng.annotations.Test;
 import static org.testng.Assert.*;
 
 import com.datastax.driver.core.policies.*;
+import com.datastax.driver.core.exceptions.NoHostAvailableException;
 import static com.datastax.driver.core.TestUtils.*;
 
 public class LoadBalancingPolicyTest {
@@ -165,6 +166,93 @@ public class LoadBalancingPolicyTest {
     @Test(groups = "integration")
     public void tokenAwarePreparedTest() throws Throwable {
         tokenAwareTest(true);
+    }
+
+    @Test(groups = "integration")
+    public void dcAwareRoundRobinTestWithOneRemoteHost() throws Throwable {
+
+        Cluster.Builder builder = Cluster.builder().withLoadBalancingPolicy(new DCAwareRoundRobinPolicy("dc2", 1));
+        CCMBridge.CCMCluster c = CCMBridge.buildCluster(2, 2, builder);
+        createMultiDCSchema(c.session);
+        try {
+
+            init(c, 12);
+            query(c, 12);
+
+            assertQueried(CCMBridge.IP_PREFIX + "1", 0);
+            assertQueried(CCMBridge.IP_PREFIX + "2", 0);
+            assertQueried(CCMBridge.IP_PREFIX + "3", 6);
+            assertQueried(CCMBridge.IP_PREFIX + "4", 6);
+            assertQueried(CCMBridge.IP_PREFIX + "5", 0);
+
+            resetCoordinators();
+            c.cassandraCluster.bootstrapNode(5, "dc3");
+            waitFor(CCMBridge.IP_PREFIX + "5", c.cluster, 20);
+
+            query(c, 12);
+
+            assertQueried(CCMBridge.IP_PREFIX + "1", 0);
+            assertQueried(CCMBridge.IP_PREFIX + "2", 0);
+            assertQueried(CCMBridge.IP_PREFIX + "3", 6);
+            assertQueried(CCMBridge.IP_PREFIX + "4", 6);
+            assertQueried(CCMBridge.IP_PREFIX + "5", 0);
+
+            resetCoordinators();
+            c.cassandraCluster.decommissionNode(3);
+            c.cassandraCluster.decommissionNode(4);
+            waitForDecommission(CCMBridge.IP_PREFIX + "3", c.cluster, 20);
+            waitForDecommission(CCMBridge.IP_PREFIX + "4", c.cluster, 20);
+
+            query(c, 12);
+
+            assertQueried(CCMBridge.IP_PREFIX + "1", 0);
+            assertQueried(CCMBridge.IP_PREFIX + "2", 0);
+            assertQueried(CCMBridge.IP_PREFIX + "3", 0);
+            assertQueried(CCMBridge.IP_PREFIX + "4", 0);
+            assertQueried(CCMBridge.IP_PREFIX + "5", 12);
+
+            resetCoordinators();
+            c.cassandraCluster.decommissionNode(5);
+            waitForDecommission(CCMBridge.IP_PREFIX + "5", c.cluster, 20);
+
+            query(c, 12);
+
+            assertQueried(CCMBridge.IP_PREFIX + "1", 0);
+            assertQueried(CCMBridge.IP_PREFIX + "2", 12);
+            assertQueried(CCMBridge.IP_PREFIX + "3", 0);
+            assertQueried(CCMBridge.IP_PREFIX + "4", 0);
+            assertQueried(CCMBridge.IP_PREFIX + "5", 0);
+
+            resetCoordinators();
+            c.cassandraCluster.decommissionNode(2);
+            waitForDecommission(CCMBridge.IP_PREFIX + "2", c.cluster, 20);
+
+            query(c, 12);
+
+            assertQueried(CCMBridge.IP_PREFIX + "1", 12);
+            assertQueried(CCMBridge.IP_PREFIX + "2", 0);
+            assertQueried(CCMBridge.IP_PREFIX + "3", 0);
+            assertQueried(CCMBridge.IP_PREFIX + "4", 0);
+            assertQueried(CCMBridge.IP_PREFIX + "5", 0);
+
+            resetCoordinators();
+            c.cassandraCluster.forceStop(1);
+
+            // TODO: This should throw an error and be caught
+            try {
+                query(c, 12);
+                assertTrue(false);
+            } catch (NoHostAvailableException e) {
+                // No more nodes so ...
+            }
+
+        } catch (Throwable e) {
+            c.errorOut();
+            throw e;
+        } finally {
+            resetCoordinators();
+            c.discard();
+        }
     }
 
     public void tokenAwareTest(boolean usePrepared) throws Throwable {